Expected effects
The degree and rate of physical effects is dependent on the dose and route of administration, as well as client-specific factors such as age, genetics, body habitus and lifestyle.

Desired androgenic effects of testosterone therapy include deepened voice, cessation of menses, clitoral growth, increased muscle mass, and hair growth in androgen dependent areas including facial hair. Typically, clients taking testosterone develop a male phenotype over a period of months to years. The timeframe of physiologic changes may be slightly slower with the use of transdermal preparations.

! Hormone therapy results in both reversible and irreversible masculinization.