Consent form for masculinizing hormone therapy
As with any other medical intervention, clients must demonstrate an understanding of the risks and benefits of hormone treatment. Obtaining informed consent is a process that primary care providers engage in daily, and when prescribing hormones to trans clients, the same basic principles apply. Questions may arise around capacity to consent in individuals with cognitive or development disabilities, mental health and substance use challenges, or younger clients. Sherbourne Health Centre implements an ‘informed consent model’ for hormone provision. In this model, the focus is on obtaining informed consent as the threshold for the initiation of hormone therapy, with less emphasis on meeting DSM diagnostic criteria for Gender Dysphoria or requiring a mental health assessment unless significant mental health concerns are identified.